Revert "TP02 - Relaciones MedioPago y Factura y DTO implementados (#78)"
This reverts commit baa6d939563ee39245daab1f3c7b8b0d1dc285ff.
