+++ src/main/java/com/imb2025/smedico/controller/FacturaController.java
@@ -1,6 +1,5 @@
 package com.imb2025.smedico.controller;
 
-import com.imb2025.smedico.dto.FacturaRequestDTO;
 import com.imb2025.smedico.entity.Factura;
 import com.imb2025.smedico.service.IFacturaService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,35 +25,13 @@ public class FacturaController {
     }
 
     @PostMapping
-    public Factura createFactura(@RequestBody FacturaRequestDTO requestDTO){
-        try {
-            return facturaService.create(facturaService.fromDto(requestDTO));
-        } catch (Exception e){
-            e.printStackTrace();
-            return null;
-        }
+    public Factura createFactura(@RequestBody Factura factura){
+        return facturaService.save(factura);
     }
 
-    /*
-    @PostMapping
-    public Factura crearFactura(@RequestBody FacturaRequestDTO dto) {
-        try {
-            return facturaService.createFactura(dto); //Me susbraya en rojo en createFactura
-        } catch (Exception e) {
-            System.out.println("Error al crear Factura: " + e.getMessage());
-            return null;
-        }
-    }
-    */
-    
-    @PutMapping("/{id}")
-    public Factura updateFactura(@PathVariable Long id, @RequestBody FacturaRequestDTO requestDTO){
-        try{
-            return facturaService.update(id, facturaService.fromDto(requestDTO));
-        } catch (Exception e){
-            e.printStackTrace();
-            return null;
-        }
+    @PutMapping
+    public Factura updateFactura(@RequestBody Factura factura){
+        return facturaService.save(factura);
     }
 
     @DeleteMapping("/{id}")
