+++ src/main/java/com/imb2025/smedico/entity/Factura.java
@@ -1,6 +1,9 @@
 package com.imb2025.smedico.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 
 import java.time.LocalDate;
 
@@ -13,28 +16,11 @@ public class Factura {
 
     private LocalDate fecha;
 
+    private Long pacienteId;
+
     private Double monto;
 
-    @ManyToOne
-    @JoinColumn(name = "medio_pago_id")
-    private MedioPago medioPago; // Cambiado: Usamos la entidad en lugar del Id
-
-    public MedioPago getMedioPago() { //Agregado
-		return medioPago;
-	}
-
-	public void setMedioPago(MedioPago medioPago) { //Agregado
-		this.medioPago = medioPago;
-	}
-
-	public Factura() {
-    }
-
-    public Factura(LocalDate fecha, Double monto, MedioPago medioPago) {
-        this.fecha = fecha;
-        this.monto = monto;
-        this.medioPago = medioPago; //Modificado: dato tipo Entidad MedioPago
-    }
+    private Long medioPagoId;
 
     public Long getId() {
         return id;
@@ -52,6 +38,14 @@ public class Factura {
         this.fecha = fecha;
     }
 
+    public Long getPacienteId() {
+        return pacienteId;
+    }
+
+    public void setPacienteId(Long pacienteId) {
+        this.pacienteId = pacienteId;
+    }
+
     public Double getMonto() {
         return monto;
     }
@@ -60,5 +54,11 @@ public class Factura {
         this.monto = monto;
     }
 
-    
+    public Long getMedioPagoId() {
+        return medioPagoId;
+    }
+
+    public void setMedioPagoId(Long medioPagoId) {
+        this.medioPagoId = medioPagoId;
+    }
 }
