+++ src/main/java/com/imb2025/smedico/entity/MedioPago.java
@@ -1,15 +1,10 @@
 package com.imb2025.smedico.entity;
 
-import java.util.List;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 
 @Entity
 public class MedioPago {
@@ -18,25 +13,11 @@ public class MedioPago {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String nombre;
-	
-	@Enumerated(EnumType.STRING)
-    //@Column(nullable = false)
+	//@Enumerated(EnumType.STRING)
+    @Column(nullable = false)
 	private TipoPago tipo;
 	
-	@OneToMany(mappedBy = "medioPago")
-    private List<Factura> facturas; // nos relacionamos con Factura
 	
-	
-	public List<Factura> getFacturas() {
-		return facturas;
-	}
-
-
-	public void setFacturas(List<Factura> facturas) {
-		this.facturas = facturas;
-	}
-
-
 	public Long getId() {
 		return id;
 	}
