+++ src/main/java/com/imb2025/smedico/service/IMedioPagoService.java
@@ -1,9 +1,9 @@
 package com.imb2025.smedico.service;
 
 import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import com.imb2025.smedico.dto.MedioPagoRequestDTO;
 import com.imb2025.smedico.entity.MedioPago;
 
 @Service
@@ -12,10 +12,5 @@ public interface IMedioPagoService {
 	public MedioPago findById(Long id);
     public MedioPago save(MedioPago medioPago);
     public MedioPago update(Long id, MedioPago medioPago);
-    public void deleteById(Long id);
-    
-	public MedioPago createMedioPago(MedioPagoRequestDTO dto);  
-    
-
-	
+    public void deleteById(Long id);  
 }
